Move contacts modal UI into app layer
